1. Finished implementing the LightArmor class with different decorators. 2. Corrected repeat variables in IItem and IEquipment. 3. Implemented string formating for the weapon stats display.

+++ Model/Gameplay/Items/Interfaces/IItem.cs
@@ -24,8 +24,8 @@ namespace ShaRPGame.Model.GameplayModels.Items.ItemInterfaces
         double GetItemWeight();
         void SetItemWeight(double value);
 
-        double GetItemRarity();
-        void SetItemRarity(double value);
+        int GetItemRarity();
+        void SetItemRarity(int value);
 
         public int StatBonus(CharacterModel character);
     }
